수정

+++ src/org/iclass/vo/Community.java
@@ -14,7 +14,7 @@ import lombok.ToString;
 @Builder
 @ToString
 public class Community {
-	private long idx;
+	private int idx;
 	private String writer;
 	private String title;
 	private String content;
